requirements.txt
tensorflow==2.8.0
numpy==1.20.2
Pillow==8.1.0
requests==2.25.1
protobuf==3.14.0
streamlit==0.79.0
